backend/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
firebase-admin==6.6.0
google-cloud-speech==2.27.0
openai==1.51.2
stripe==10.7.0
pydantic==2.9.2
pydantic-settings==2.4.0
python-dotenv==1.0.1
httpx==0.27.2
sentry-sdk==2.13.0
pytest==8.3.2
pytest-asyncio==0.23.7
ruff==0.6.9
black==24.8.0